Follow a more traditional R package setup for data
Raw data are in /inst/extdata as CSV files.

All derived data are in /data and lazyloaded so all you need to do is call `library("ChickpeaAscoDispersal")` and the cleaned weather data, summarised lesion counts and summarised weather data are available in the R session.

This greatly cleans up the vignettes and enhances the package structure.